create header for forward declarations of the various point types and move relevant things from coordinates.h to it
