+++ src/coordinates.h
@@ -12,6 +12,7 @@
 #include <vector>
 
 #include "coordinate_conversions.h"
+#include "coords_fwd.h"
 #include "cuboid_rectangle.h"
 #include "debug.h"
 #include "game_constants.h"
@@ -25,21 +26,6 @@ enum class direction : unsigned;
 
 namespace coords
 {
-
-enum class scale {
-    map_square,
-    submap,
-    overmap_terrain,
-    segment,
-    overmap,
-    vehicle
-};
-
-constexpr scale ms = scale::map_square;
-constexpr scale sm = scale::submap;
-constexpr scale omt = scale::overmap_terrain;
-constexpr scale seg = scale::segment;
-constexpr scale om = scale::overmap;
 
 constexpr int map_squares_per( scale s )
 {
@@ -61,15 +47,6 @@ constexpr int map_squares_per( scale s )
             constexpr_fatal( 0, "Requested scale of %d", s );
     }
 }
-
-enum class origin {
-    relative, // this is a special origin that can be added to any other
-    abs, // the global absolute origin for the entire game
-    submap, // from corner of submap
-    overmap_terrain, // from corner of overmap_terrain
-    overmap, // from corner of overmap
-    reality_bubble, // from corner of a reality bubble (aka 'map' or 'tinymap')
-};
 
 constexpr origin origin_from_scale( scale s )
 {
@@ -224,7 +201,7 @@ class coord_point_mut< Point, Subpoint, true> : public coord_point_base<Point>
         constexpr coord_point_mut( T x, T y, T z ) : base( x, y, z ) {}
 };
 
-template<typename Point, origin Origin, scale Scale, bool InBounds = false>
+template<typename Point, origin Origin, scale Scale, bool InBounds>
 class coord_point : public
     coord_point_mut<Point, coord_point<point, Origin, Scale, InBounds>, InBounds>
 {
@@ -671,63 +648,6 @@ struct std::hash<coords::coord_point<Point, Origin, Scale, InBounds>> {
         return h( p.raw() );
     }
 };
-
-/** Typedefs for point types with coordinate mnemonics.
- *
- * Each name is of the form (tri)point_<origin>_<scale>(_ib) where <origin> tells you
- * the context in which the point has meaning, and <scale> tells you what one unit
- * of the point means. The optional "_ib" suffix denotes that the type is guaranteed
- * to be inbounds.
- *
- * For example:
- * point_omt_ms is the position of a map square within an overmap terrain.
- * tripoint_rel_sm is a relative tripoint submap offset.
- *
- * For more details see doc/POINTS_COORDINATES.md.
- */
-/*@{*/
-using point_rel_ms = coords::coord_point<point, coords::origin::relative, coords::ms>;
-using point_abs_ms = coords::coord_point<point, coords::origin::abs, coords::ms>;
-using point_sm_ms = coords::coord_point<point, coords::origin::submap, coords::ms>;
-using point_sm_ms_ib = coords::coord_point<point, coords::origin::submap, coords::ms, true>;
-using point_omt_ms = coords::coord_point<point, coords::origin::overmap_terrain, coords::ms>;
-using point_bub_ms = coords::coord_point<point, coords::origin::reality_bubble, coords::ms>;
-using point_bub_ms_ib =
-    coords::coord_point<point, coords::origin::reality_bubble, coords::ms, true>;
-using point_rel_sm = coords::coord_point<point, coords::origin::relative, coords::sm>;
-using point_abs_sm = coords::coord_point<point, coords::origin::abs, coords::sm>;
-using point_omt_sm = coords::coord_point<point, coords::origin::overmap_terrain, coords::sm>;
-using point_om_sm = coords::coord_point<point, coords::origin::overmap, coords::sm>;
-using point_bub_sm = coords::coord_point<point, coords::origin::reality_bubble, coords::sm>;
-using point_bub_sm_ib =
-    coords::coord_point<point, coords::origin::reality_bubble, coords::sm, true>;
-using point_rel_omt = coords::coord_point<point, coords::origin::relative, coords::omt>;
-using point_abs_omt = coords::coord_point<point, coords::origin::abs, coords::omt>;
-using point_om_omt = coords::coord_point<point, coords::origin::overmap, coords::omt>;
-using point_abs_seg = coords::coord_point<point, coords::origin::abs, coords::seg>;
-using point_rel_om = coords::coord_point<point, coords::origin::relative, coords::om>;
-using point_abs_om = coords::coord_point<point, coords::origin::abs, coords::om>;
-
-using tripoint_rel_ms = coords::coord_point<tripoint, coords::origin::relative, coords::ms>;
-using tripoint_abs_ms = coords::coord_point<tripoint, coords::origin::abs, coords::ms>;
-using tripoint_sm_ms = coords::coord_point<tripoint, coords::origin::submap, coords::ms>;
-using tripoint_sm_ms_ib = coords::coord_point<tripoint, coords::origin::submap, coords::ms, true>;
-using tripoint_omt_ms = coords::coord_point<tripoint, coords::origin::overmap_terrain, coords::ms>;
-using tripoint_bub_ms = coords::coord_point<tripoint, coords::origin::reality_bubble, coords::ms>;
-using tripoint_bub_ms_ib =
-    coords::coord_point<tripoint, coords::origin::reality_bubble, coords::ms, true>;
-using tripoint_rel_sm = coords::coord_point<tripoint, coords::origin::relative, coords::sm>;
-using tripoint_abs_sm = coords::coord_point<tripoint, coords::origin::abs, coords::sm>;
-using tripoint_om_sm = coords::coord_point<tripoint, coords::origin::overmap, coords::sm>;
-using tripoint_bub_sm = coords::coord_point<tripoint, coords::origin::reality_bubble, coords::sm>;
-using tripoint_bub_sm_ib =
-    coords::coord_point<tripoint, coords::origin::reality_bubble, coords::sm, true>;
-using tripoint_rel_omt = coords::coord_point<tripoint, coords::origin::relative, coords::omt>;
-using tripoint_abs_omt = coords::coord_point<tripoint, coords::origin::abs, coords::omt>;
-using tripoint_om_omt = coords::coord_point<tripoint, coords::origin::overmap, coords::omt>;
-using tripoint_abs_seg = coords::coord_point<tripoint, coords::origin::abs, coords::seg>;
-using tripoint_abs_om = coords::coord_point<tripoint, coords::origin::abs, coords::om>;
-/*@}*/
 
 using coords::project_to;
 using coords::project_remain;
